Add repositories, populate

+++ PointOfSaleSystem.API/RequestBodies/Tax/UpdateTaxRequest.cs
@@ -5,7 +5,7 @@ namespace PointOfSaleSystem.API.RequestBodies.Tax
     public class UpdateTaxRequest
     {
         public required Guid Id { get; set; }
-        public required float amount { get; set; }
+        public required decimal amount { get; set; }
         public required DateTime UpdateTime { get; set; }
     }
 }
